[experimental/uniad] Eliminate host roundtrips on the UniAD warm path
Each UniAD forward used to pull intermediate tensors back to host
multiple times per attention layer / head to do simple arithmetic or
indexing in torch, then re-upload — every round-trip is a PCI-e
transfer that blocks the device. This commit replaces those patterns
with pure-ttnn equivalents across the warm path, plus a few structural
fixes needed to keep the model multi-call safe.

Modules touched:

- Decoder: `ttnn_deformable_attention.py`, `ttnn_detr_transformer_decoder*.py`
  Grid normalization and inverse-sigmoid now done with ttnn divide /
  subtract / rsub (rsub replaces an `ttnn.ones(shape=x.shape) - x` that
  also blocked trace capture). Spatial-shape Python ints cached on the
  attention instance so .item() reads happen during warm-up, not under
  trace.

- BEVFormer encoder: `ttnn_encoder.py`, `ttnn_temporal_self_attention.py`,
  `ttnn_spatial_cross_attention.py`
  Per-forward precompute of valid-query indexes, the count tensor, the
  scatter-add index tensors, and the rebatched reference_points_cam
  tensor — all of which are bev_mask-derived and identical across
  the 6 encoder layers. The per-layer body now consumes these via
  kwargs instead of recomputing them, and the result accumulation step
  uses ttnn.scatter_add on device instead of pulling queries back to
  host. Adds a single-gather rebatch path (`_precomputed_gather_index_concat`)
  for queries.

- Motion / pan-seg / planning / query-interaction / detr-track-3d-coder:
  pure-ttnn slice + concat replaces torch slice-assignment round-trips.
  `ttnn_uniad.py` track_scores: pure ttnn.max replaces
  `to_torch(...).max(dim=-1).values`.

- prev_bev != None path in TtPerceptionTransformer: implement the
  rotation branch end-to-end (host torchvision.rotate + reupload — no
  ttnn rotate op yet, but the surrounding plumbing is correct).

- `ttnn_utils.py`:
  - Instances.cat: concat on dim=0 (instance axis), matching torch.cat.
    The previous dim=1 was invalid for rank-1 fields like obj_idxes.
  - Instances.__len__: handle ttnn.Tensor fields by returning shape[0].
    Required for the second-call path's `len(active_inst) > 0` check.
  - Instances getitem: promote UInt32 round-trip indexes to torch.long
    so torch CPU indexing accepts them.

- Diagnostic harnesses (env-gated, inert by default):
  - TT_UNIAD_TIMING=1 in `ttnn_uniad.py` — print per-sub-phase
    wall-clock breakdown of the forward call (skips Call #1 / cold).
  - TT_DCN_TIMING=1 in `ttnn_resnet.py` — accumulate transfer / CPU /
    back time across the 26 modulated-deformable-conv calls inside
    TtModulatedDeformConv2dPack and print on every ResNet forward.
    This is what surfaced that the CPU `modulated_deform_conv_forward`
    body (not the PCI-e transfers around it) is the dominant cost of
    the DCN path.

+++ models/experimental/uniad/tt/ttnn_seg_deformable_transformer.py
@@ -98,11 +98,14 @@ class TtSegDeformableTransformer:
         mask_flatten = ttnn.concat(mask_flatten, 1)
         lvl_pos_embed_flatten = ttnn.concat(lvl_pos_embed_flatten, 1)
 
+        # ttnn.prod requires BFLOAT16 (prod_nc_device_operation.cpp:37); the
+        # spatial_shapes tensor only holds small shape integers, so bfloat16
+        # carries no meaningful precision loss.
         spatial_shapes = ttnn.from_torch(
             torch.as_tensor(spatial_shapes, dtype=torch.long),
             device=self.device,
             layout=ttnn.TILE_LAYOUT,
-            dtype=ttnn.float32,
+            dtype=ttnn.bfloat16,
         )
         prod = ttnn.prod(spatial_shapes, dim=1)
         cumsum = ttnn.cumsum(prod, dim=0)
